modularize localStorage calls into localStorage.js in preparation for unit testing
